Made usability updates and comment removals.
I removed the comments that describe the keyboard layout because they are redundant and I don't want to maintain them.

Underscore can now be typed while holding shift instead of lower. This is more comfortable to type since the thumb position for shift is more comfortable to me. It also just feels like it makes sense to have underscore be what's typed when shift space is pressed. I believe relying on "feels right" is quite important when it comes to something very personal like this layout.

Backspace is now a lone key that does the same thing no matter the layer. I already encountered a situation where I pressed backspace, but it did something else because I was on a different layer. I very quickly realized how annoying that would be, so I moved the insert and print screen commands down to where the escape key is. This means these "command type" keys can mostly be found on the r21 switch and backspace can be a dedicated key.

Control moved to switch l19, where lower used to be. CapsLock came along with it. I realized that I need to keep control accessible to the left hand. That way I can still press control combos with only my left hand. That's pretty important when I am going through workflows that require both the keyboard and mouse simultaneously. Lower is also not as imporant of a key, so it makes sense for it to take up a less comfortable position.