Modified command key configurations
I was trying to lock the computer with the GUI+l and it didn't work because I was holding the raise key while pressing l. I can lock the computer by letting go of the raise key before pressing l, but I find this to be uncomfortable for a key combo that I use so often. It is furthermore clear to me that GUI should have its own key because it makes a lot of key combos easier and sometimes even possible. Take GUI+SHFT+s for example (screen clip). This was, if not very hard, impossible with the last scheme, but now becomes possible with just the left hand.

I swapped CAPS and RET to have access to SHFT+RET. It's still not the most comfortable thing, but it's a rarely used key combo. I only use it when sending messages that I instert a newline into. It also just makes a lot of sense to have CAPS be on the raised layer of the SHFT key and enter to be on the raised layer of the CTRL key.

With GUI gone from rrse21, room was left for the missing delete key that I need in some programs.