Rotated r11->r17->r12
I notice my right pinky continue to reach for the where it thinks '/' should be. I am deciding to cater to its desire while also catering to my desire to have single quote and double quote directly under my right pinky. The movement of ':' is just a consequence of this decision.